Migrate missing routines to GPXMultiParserWPT
