+++ main/src/main/java/cgeo/geocaching/files/GPXMultiParserWPT.java
@@ -5,17 +5,24 @@ import cgeo.geocaching.connector.IConnector;
 import cgeo.geocaching.connector.capability.ILogin;
 import cgeo.geocaching.connector.gc.GCConnector;
 import cgeo.geocaching.connector.gc.GCUtils;
+import cgeo.geocaching.connector.internal.InternalConnector;
 import cgeo.geocaching.connector.tc.TerraCachingLogType;
 import cgeo.geocaching.connector.tc.TerraCachingType;
 import cgeo.geocaching.enumerations.CacheAttribute;
 import cgeo.geocaching.enumerations.CacheSize;
 import cgeo.geocaching.enumerations.CacheType;
+import cgeo.geocaching.enumerations.LoadFlags;
+import cgeo.geocaching.enumerations.WaypointType;
+import cgeo.geocaching.list.StoredList;
 import cgeo.geocaching.location.Geopoint;
 import cgeo.geocaching.log.LogEntry;
 import cgeo.geocaching.log.LogType;
 import cgeo.geocaching.models.Geocache;
-import cgeo.geocaching.models.INamedGeoCoordinate;
 import cgeo.geocaching.models.Trackable;
+import cgeo.geocaching.models.Waypoint;
+import cgeo.geocaching.models.WaypointUserNoteCombiner;
+import cgeo.geocaching.storage.DataStore;
+import cgeo.geocaching.utils.DisposableHandler;
 import cgeo.geocaching.utils.Log;
 import cgeo.geocaching.utils.MatcherWrapper;
 import cgeo.geocaching.utils.SynchronizedDateFormat;
@@ -31,8 +38,12 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Date;
+import java.util.EnumSet;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Locale;
+import java.util.Set;
+import java.util.concurrent.CancellationException;
 import java.util.regex.Pattern;
 
 import org.apache.commons.lang3.StringUtils;
@@ -53,7 +64,6 @@ public class GPXMultiParserWPT extends GPXMultiParserBase {
 
     private static final Pattern PATTERN_MILLISECONDS = Pattern.compile("\\.\\d{3,7}");
 
-    private final ArrayList<INamedGeoCoordinate> result = new ArrayList<>();
     private final int listId;
 
     private Geocache cache;
@@ -75,6 +85,12 @@ public class GPXMultiParserWPT extends GPXMultiParserBase {
     private boolean wptEmptyCoordinates = false;
     private int cacheAssignedEmoji = 0;
     private List<LogEntry> logs = new ArrayList<>();
+
+    /**
+     * Parser result. Maps geocode to cache.
+     */
+    private final Set<String> result = new HashSet<>(100);
+    private ProgressInputStream progressStream;
 
     /**
      * original longitude in case of modified coordinates
@@ -195,6 +211,213 @@ public class GPXMultiParserWPT extends GPXMultiParserBase {
         registerExtensions(node.get("extensions"));
         registerExtensions(node);
 
+        // former waypoint.endElementListener -----------------------------------------------------
+
+        // try to find geocode somewhere else
+        if (StringUtils.isBlank(cache.getGeocode())) {
+            findGeoCode(name, true);
+            findGeoCode(desc, false);
+            findGeoCode(cmt, false);
+        }
+
+        // take the name as code, if nothing else is available
+        if (StringUtils.isBlank(cache.getGeocode()) && StringUtils.isNotBlank(name)) {
+            cache.setGeocode(name.trim());
+        }
+
+        if (isValidForImport()) {
+            fixCache(cache);
+            if (listId != StoredList.TEMPORARY_LIST.id) {
+                cache.getLists().add(listId);
+            }
+            cache.setDetailed(true);
+            cache.setLogPasswordRequired(logPasswordRequired);
+            if (StringUtils.isNotBlank(descriptionPrefix)) {
+                cache.setDescription(descriptionPrefix + cache.getDescription());
+            }
+
+            createNoteFromGSAKUserdata();
+
+            cache.setAssignedEmoji(cacheAssignedEmoji);
+
+            final String geocode = cache.getGeocode();
+            if (result.contains(geocode)) {
+                Log.w("Duplicate geocode during GPX import: " + geocode);
+            }
+            // modify cache depending on the use case/connector
+            afterParsing(cache);
+
+            // finally store the cache in the database
+            result.add(geocode);
+            DataStore.saveCache(cache, EnumSet.of(LoadFlags.SaveFlag.DB));
+            DataStore.saveLogs(cache.getGeocode(), logs, false);
+
+            // avoid the cachecache using lots of memory for caches which the user did not actually look at
+            DataStore.removeCache(geocode, EnumSet.of(LoadFlags.RemoveFlag.CACHE));
+//todo            showProgressMessage(progressHandler, progressStream.getProgress());
+        } else if (StringUtils.isNotBlank(cache.getName())
+                && (StringUtils.containsIgnoreCase(type, "waypoint") || terraChildWaypoint)) {
+            addWaypointToCache();
+        }
+
+        resetCache();
+
+    }
+
+    /** create a cache note from the UserData1 to UserData4 fields supported by GSAK */
+    private void createNoteFromGSAKUserdata() {
+        if (StringUtils.isBlank(cache.getPersonalNote())) {
+            final StringBuilder buffer = new StringBuilder();
+            for (final String anUserData : userData) {
+                if (StringUtils.isNotBlank(anUserData)) {
+                    buffer.append(' ').append(anUserData);
+                }
+            }
+            final String note = buffer.toString().trim();
+            if (StringUtils.isNotBlank(note)) {
+                cache.setPersonalNote(note, true);
+            }
+        }
+    }
+
+    private boolean isValidForImport() {
+        final String geocode = cache.getGeocode();
+        if (StringUtils.isBlank(geocode)) {
+            return false;
+        }
+
+        final boolean isInternal = InternalConnector.getInstance().canHandle(geocode);
+        if (cache.getCoords() == null && !isInternal) {
+            return false;
+        }
+        final boolean valid = (type == null && subtype == null && sym == null)
+                || StringUtils.contains(type, "geocache")
+                || StringUtils.contains(sym, "geocache")
+                || StringUtils.containsIgnoreCase(sym, "waymark")
+                || (StringUtils.containsIgnoreCase(sym, "terracache") && !terraChildWaypoint);
+        if ("GC_WayPoint1".equals(cache.getShortDescription())) {
+            terraChildWaypoint = true;
+        }
+        return valid;
+    }
+
+    /**
+     * Overwrite this method in a GPX parser sub class to modify the {@link Geocache}, after it has been fully parsed
+     * from the GPX file and before it gets stored.
+     *
+     * @param cache currently imported cache
+     */
+    protected void afterParsing(final Geocache cache) {
+        if ("GC_WayPoint1".equals(cache.getShortDescription())) {
+            cache.setShortDescription("");
+        }
+    }
+
+    private void addWaypointToCache() {
+        fixCache(cache);
+
+        if (cache.getName().length() > 2 || StringUtils.isNotBlank(parentCacheCode)) {
+            if (StringUtils.isBlank(parentCacheCode)) {
+                if (StringUtils.containsIgnoreCase(scriptUrl, "extremcaching")) {
+                    parentCacheCode = cache.getName().substring(2);
+                } else if (terraChildWaypoint) {
+                    parentCacheCode = StringUtils.left(cache.getGeocode(), cache.getGeocode().length() - 1);
+                } else {
+                    parentCacheCode = "GC" + cache.getName().substring(2).toUpperCase(Locale.US);
+                }
+            }
+
+            if ("GC_WayPoint1".equals(cache.getShortDescription())) {
+                cache.setShortDescription("");
+            }
+
+            final Geocache cacheForWaypoint = findParentCache();
+            if (cacheForWaypoint != null) {
+                final Waypoint waypoint = new Waypoint(cache.getShortDescription(), WaypointType.fromGPXString(sym, subtype), false);
+                if (wptUserDefined) {
+                    waypoint.setUserDefined();
+                }
+                waypoint.setId(-1);
+                waypoint.setGeocode(parentCacheCode);
+                String cacheName = cache.getName();
+                if (wptUserDefined) {
+                    // try to deduct original prefix from wpt name
+                    if (StringUtils.endsWithIgnoreCase(cacheName, parentCacheCode.substring(2))) {
+                        cacheName = cacheName.substring(0, cacheName.length() - parentCacheCode.length() + 2);
+                    }
+                    if (StringUtils.startsWithIgnoreCase(cacheName, Waypoint.PREFIX_OWN + "-")) {
+                        cacheName = cacheName.substring(4);
+                    }
+                }
+                waypoint.setPrefix(cacheForWaypoint.getWaypointPrefix(cacheName));
+                waypoint.setLookup("---");
+                // there is no lookup code in gpx file
+
+                waypoint.setCoords(cache.getCoords());
+
+                // set flag for user-modified coordinates of cache
+                if (waypoint.getWaypointType() == WaypointType.ORIGINAL) {
+                    cacheForWaypoint.setUserModifiedCoords(true);
+                }
+
+                // user defined waypoint does not have original empty coordinates
+                if (wptEmptyCoordinates || (!waypoint.isUserDefined() && null == waypoint.getCoords())) {
+                    waypoint.setOriginalCoordsEmpty(true);
+                }
+
+                final WaypointUserNoteCombiner wpCombiner = new WaypointUserNoteCombiner(waypoint);
+                wpCombiner.updateNoteAndUserNote(cache.getDescription());
+
+                waypoint.setVisited(wptVisited);
+                final List<Waypoint> mergedWayPoints = new ArrayList<>(cacheForWaypoint.getWaypoints());
+
+                final List<Waypoint> newPoints = new ArrayList<>();
+                newPoints.add(waypoint);
+                Waypoint.mergeWayPoints(newPoints, mergedWayPoints, true);
+                cacheForWaypoint.setWaypoints(newPoints, false);
+                DataStore.saveCache(cacheForWaypoint, EnumSet.of(LoadFlags.SaveFlag.DB));
+//todo                showProgressMessage(progressHandler, progressStream.getProgress());
+            }
+        }
+    }
+
+    @Nullable
+    private Geocache findParentCache() {
+        if (StringUtils.isBlank(parentCacheCode)) {
+            return null;
+        }
+        // first match by geocode only
+        Geocache cacheForWaypoint = DataStore.loadCache(parentCacheCode, LoadFlags.LOAD_CACHE_OR_DB);
+        if (cacheForWaypoint == null) {
+            // then match by title
+            final String geocode = DataStore.getGeocodeForTitle(parentCacheCode);
+            if (StringUtils.isNotBlank(geocode)) {
+                cacheForWaypoint = DataStore.loadCache(geocode, LoadFlags.LOAD_CACHE_OR_DB);
+            }
+        }
+        return cacheForWaypoint;
+    }
+
+    protected void showProgressMessage(@Nullable final DisposableHandler handler, final int bytesRead) {
+        if (handler != null) {
+            if (handler.isDisposed()) {
+                throw new CancellationException();
+            }
+            handler.sendMessage(handler.obtainMessage(0, bytesRead, 0));
+        }
+    }
+
+    protected void fixCache(final Geocache cache) {
+        final List<Trackable> inventory = cache.getInventory();
+        cache.setInventoryItems(inventory.size());
+        final long time = System.currentTimeMillis();
+        cache.setUpdated(time);
+        cache.setDetailedUpdate(time);
+
+        // fix potentially bad cache id
+        if (GCConnector.getInstance().equals(ConnectorFactory.getConnector(cache))) {
+            cache.setCacheId(String.valueOf(GCUtils.gcLikeCodeToGcLikeId(cache.getGeocode())));
+        }
     }
 
     private void registerExtensions(@Nullable final XmlNode cacheParent) {
@@ -380,13 +603,11 @@ public class GPXMultiParserWPT extends GPXMultiParserBase {
         }
 
         gsak.onNonBlankValue("Watch", value -> cache.setOnWatchlist(Boolean.parseBoolean(value.trim())));
-/* todo
-        gsak.getChild(gsakNamespace, "UserData").setEndTextElementListener(new GPXMultiParserCaches.UserDataListener(1));
-
+        gsak.onNonBlankValue("UserData", value -> userData[1] = value);
         for (int i = 2; i <= 4; i++) {
-            gsak.getChild(gsakNamespace, "User" + i).setEndTextElementListener(new GPXMultiParserCaches.UserDataListener(i));
+            final int finalI = i;
+            gsak.onNonBlankValue("User" + i, value -> userData[finalI] = value);
         }
-*/
         gsak.onNonBlankValue("Parent", value -> parentCacheCode = value);
         gsak.onDefinedIntegerValue("FavPoints", value -> cache.setFavoritePoints(value));
         gsak.onNonBlankValue("GcNote", value -> cache.setPersonalNote(StringUtils.trim(value), true));
@@ -521,17 +742,10 @@ public class GPXMultiParserWPT extends GPXMultiParserBase {
         });
     }
 
-
-
-
-
     @Override
     void onParsingDone(@NonNull final Collection<Object> result) {
-        result.addAll(this.result);
+        result.addAll(DataStore.loadCaches(this.result, EnumSet.of(LoadFlags.LoadFlag.DB_MINIMAL)));
     }
-
-
-
 
     protected void setUrl(final String url) {
         // try to find guid somewhere else
